backend/requirements.txt
fastapi
uvicorn[standard]
websockets
google-genai>=1.0.0
python-multipart
pypdf
python-docx
python-dotenv
pydantic
